Updated EditableCode to change interaction mode
