+++ frontend/src/components/layout/Page.tsx
@@ -6,7 +6,7 @@ export default function Page(
     }
 ) {
     return (
-        <div className="w-10/12 h-full pt-5 mx-auto">
+        <div className=" w-10/12 h-full pt-5 mx-auto ">
             {children}
         </div>
     )
